% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_calc.R
\name{auc_calc}
\alias{auc_calc}
\title{Calculate Area Under the Curve of glucose values for a restricted timeframe}
\usage{
auc_calc(glucose_df, timelength = 120)
}
\arguments{
\item{glucose_df}{dataframe of glucose values}

\item{timelength}{number of minutes to look ahead for the AUC calculation}
}
\description{
Returns AUC for the first timelength minutes after the start of the glucose_df
}
