% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_util.R
\name{auc_calc_components}
\alias{auc_calc_components}
\title{Helper to calculate Area Under the Curve of glucose values for a restricted timeframe}
\usage{
auc_calc_components(glucose_df, timelength = 120)
}
\arguments{
\item{glucose_df}{dataframe of glucose values}

\item{timelength}{number of minutes to look ahead for the AUC calculation}
}
\value{
a vector of the AUC values for each segment of the curve.
}
\description{
Returns each segment of the AUC for the first timelength minutes after the start of the glucose_df.
See Brouns, F., Bjorck, I., Frayn, K., Gibbs, A., Lang, V., Slama, G., & Wolever, T. (2005).
Glycaemic index methodology. Nutrition Research Reviews, 18(1), 145-171. doi:10.1079/NRR2005100
}
