% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_util.R
\name{auc_for_food}
\alias{auc_for_food}
\title{Find glucose response for all meals involving a food}
\usage{
auc_for_food(
  foodname = "Munk Pack",
  glucose_records,
  notes_records,
  prefixLength = 0,
  timeLength = 150,
  start_limit = 100
)
}
\arguments{
\item{foodname}{character string representing the food item of interest}

\item{glucose_records}{valid dataframe of glucose results}

\item{notes_records}{valid dataframe of notes}

\item{prefixLength}{number of additional minutes to add before the starttime.}

\item{timeLength}{number of minutes for the glucose record to show after the food was eaten.}

\item{start_limit}{filter to cut off any results that have more than this mg/mL when starting}
}
\value{
dataframe
}
\description{
This is a wrapper to \code{food_times_df()}. Given a foodname, returns a dataframe with every
occurence of that food, with a \code{timeLength} minutes since eating.
}
