% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cgm_data}
\alias{cgm_data}
\title{Make a Standard CGM Data Object}
\usage{
cgm_data(glucose_records, notes_records)
}
\arguments{
\item{glucose_records}{canonical glucose dataframe}

\item{notes_records}{metadata associated with the dataframe}
}
\description{
Make a Standard CGM Data Object
}
