% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_display.R
\name{cgm_display}
\alias{cgm_display}
\title{CGM display}
\usage{
cgm_display(
  start = lubridate::now() - lubridate::hours(18),
  end = lubridate::now(),
  activity_df = activity_raw,
  glucose_df = glucose_raw,
  title = "Glucose",
  show.label = TRUE
)
}
\arguments{
\item{start}{datetime for the start of the plot}

\item{end}{datetime for the end of the plot}

\item{activity_df}{activity dataframe}

\item{glucose_df}{glucose dataframe}

\item{title}{(optional) title for the plot}

\item{show.label}{(optional) show food names if TRUE}
}
\description{
show glucose levels between start and end times
}
