% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_display.R
\name{cgm_start_plot}
\alias{cgm_start_plot}
\title{Start Plot}
\usage{
cgm_start_plot(
  startTime = lubridate::now(),
  timeLength = 24,
  title = paste("Glucose", timeLength, "Hours")
)
}
\arguments{
\item{startTime}{datetime for the start of the plot}

\item{timeLength}{number of hours to display}

\item{title}{title of the plot}
}
\description{
display plot from startTime, timeLength
}
