% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_times_calc.R
\name{combined_food_times_df}
\alias{combined_food_times_df}
\title{Combine multiple glucose values}
\usage{
combined_food_times_df(food_times_df)
}
\arguments{
\item{food_times_df}{a dataframe of similar meals offset by a constant time}
}
\value{
dataframe \code{food_times_df} augmented with a new \code{ave} row
}
\description{
Given a dataframe of the form returned by \code{food_times_df()},
returns that df augmented with an additional row \code{ave}, which
represents the average across all glucose values at that time slot
}
