% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_calc.R
\name{food_auc}
\alias{food_auc}
\title{Calculate area under the curve for each item in foodlist}
\usage{
food_auc(foodlist, timelength = 120, activity_df, glucose_df)
}
\arguments{
\item{foodlist}{vector of food items}

\item{timelength}{number of minutes after the start of the food}

\item{activity_df}{activity data frame (use default activity_raw)}

\item{glucose_df}{dataframe of glucose values (use default glucose_raw)}
}
\description{
returns a dataframe giving all glucose values within "timelength" of a specific activity
}
