% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_by_meal.R
\name{food_by_meal}
\alias{food_by_meal}
\title{Process Food Results by Meal}
\usage{
food_by_meal(foodresults = NULL)
}
\arguments{
\item{foodresults}{A dataframe of food results. It should contain at least the
following three columns: 't' (numeric, representing time), 'value' (numeric, representing
the value associated with each time point), and 'meal' (character, representing the meal
associated with each time point). Typically, \code{foodresults} is a dataframe created after
calling \code{\link{food_times_df}} with a valid object created with
\code{\link{cgm_data}} and a 'foodname' string.}
}
\value{
A dataframe of the processed food results, arranged by 'meal'.
}
\description{
This function takes a food results dataframe and processes it by grouping
it by meal, arranging it by time ('t'), and adjusting the 'value' column
to start from zero for each meal. The resulting dataframe is arranged by 'meal'.
}
