% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_display.R
\name{food_compare_display}
\alias{food_compare_display}
\title{Food compare display}
\usage{
food_compare_display(
  foodlist = foodlist,
  timelength = 120,
  activity_df,
  glucose_df,
  foodtype = "food"
)
}
\arguments{
\item{foodlist}{character vector of the foods you want to plot}

\item{timelength}{time in minutes to plot}

\item{activity_df}{activity dataframe}

\item{glucose_df}{glucose dataframe}

\item{foodtype}{character vector of the label you want on the chart}
}
\description{
Plot foods compared against one another
}
