% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_calc.R
\name{food_effect}
\alias{food_effect}
\title{Calculate glucose levels for a restricted timeframe}
\usage{
food_effect(
  foodlist = c("Oatmeal", "Oatmeal w/cinnamon"),
  activity_df,
  glucose_df,
  timelength = lubridate::hours(2)
)
}
\arguments{
\item{foodlist}{vector of food items}

\item{activity_df}{activity data frame (use default activity_raw)}

\item{glucose_df}{dataframe of glucose values (use default glucose_raw)}

\item{timelength}{number of minutes after the start of the food}
}
\description{
returns a dataframe giving all glucose values within "timelength" of a specific activity
}
