% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{food_list_db}
\alias{food_list_db}
\title{List all products consumed by \code{user_id}}
\usage{
food_list_db(user_id = 1234)
}
\arguments{
\item{user_id}{vector of user IDs or NULL to show all users}
}
\value{
character vector of product names sorted alphabetically
}
\description{
List all products consumed by \code{user_id}
}
