% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_by_meal.R
\name{food_with_intervention}
\alias{food_with_intervention}
\title{Add intervention column to data frame}
\usage{
food_with_intervention(df, intervention = "glucotrojan")
}
\arguments{
\item{df}{A data frame (or tibble) to which the 'control' column will be added.}

\item{intervention}{A string to be searched for within the 'meal' column of the input data frame. Defaults to "glycotrojan".}
}
\value{
A tibble that is the same as the input tibble, but with an added 'control' column.
}
\description{
This function adds a new logical column to a tibble, named "control". The value in this column will be TRUE if
the specified intervention string is found within the 'meal' column of the data frame, and FALSE otherwise.
The string detection is case-insensitive.
}
