% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{glucose_df_for_users_at_time}
\alias{glucose_df_for_users_at_time}
\title{Glucose values for ID after startDate}
\usage{
glucose_df_for_users_at_time(
  conn_args = config::get("dataconnection"),
  user_id = 1234,
  startTime = now() - hours(36),
  timelength = 120
)
}
\arguments{
\item{conn_args}{valid database connection}

\item{user_id}{user ID}

\item{startTime}{datetime object for start time}

\item{timelength}{minutes of glucose values to return}
}
\value{
A valid glucose dataframe
}
\description{
For example \code{read_glucose_for_user_at_time(ID=22,startTime = as_datetime("2020-02-16 00:50:00", tz=Sys.timezone()))}
}
