% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{glucose_df_from_db}
\alias{glucose_df_from_db}
\title{Read from database a dataframe of glucose values for user_id ID}
\usage{
glucose_df_from_db(
  conn_args = config::get("dataconnection"),
  user_id = 1234,
  from_date = as_datetime("2000-01-01", tz = Sys.timezone()),
  db_filter = NULL
)
}
\arguments{
\item{conn_args}{valid database connection}

\item{user_id}{ID for a specific user in the database.}

\item{from_date}{a string representing the date from which you want to read the glucose values}

\item{db_filter}{A function for filtering the database. Use instead of \code{from_date} or \code{user_id}}
}
\value{
a dataframe (tibble) with the full Libreview results since fromDate for user_id
}
\description{
Reads from the current default database the glucose values for user_id ID.
}
