% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrisense_read.R
\name{glucose_df_from_levels}
\alias{glucose_df_from_levels}
\title{read a Levels export file and convert to the PSI CGM format}
\usage{
glucose_df_from_levels(
  filepath = system.file("extdata", package = "cgmr", "Firstname4Lastname4_levels.csv"),
  user_id = 3000,
  tz = "UTC"
)
}
\arguments{
\item{filepath}{path to a valid Levels CSV data file}

\item{user_id}{new user ID to be appended to the dataframe}

\item{tz}{time zone}
}
\value{
a canonical glucose value dataframe
}
\description{
Levels export files contain timestamped information about
glucose values, meals, exercise, sleep, and more.
Important: This function expects an extra line at the top of the Nutrisense raw file.
The first line must say "Levels" and include the full name of the person.
}
