% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R, R/read_data_utils.R
\name{glucose_df_from_libreview_csv}
\alias{glucose_df_from_libreview_csv}
\title{Read a standard format Libreview CSV file}
\usage{
glucose_df_from_libreview_csv(
  file = system.file("extdata", package = "psiCGM", "Firstname2Lastname2_glucose.csv"),
  user_id = 1234,
  tz = Sys.timezone()
)

glucose_df_from_libreview_csv(
  file = system.file("extdata", package = "psiCGM", "Firstname2Lastname2_glucose.csv"),
  user_id = 1234,
  tz = Sys.timezone()
)
}
\arguments{
\item{file}{path to a valid Libreview CSV file}

\item{user_id}{new user ID to be appended to the dataframe}

\item{tz}{time zone}
}
\value{
a canonical glucose value dataframe

a canonical glucose value dataframe
}
\description{
Read a valid libreview CSV file and return a dataframe and new user id
Since Libreview files don't already include a user ID, append one to the dataframe that is returned.
Importantly, datetimes are assumed to be \code{Sys.timezone()}.

Read a valid libreview CSV file and return a dataframe and new user id
Since Libreview files don't already include a user ID, append one to the dataframe that is returned.
Importantly, datetimes are assumed to be \code{Sys.timezone()}.
}
