% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{glucose_df_from_libreview_raw}
\alias{glucose_df_from_libreview_raw}
\title{Convert a Libreview Dataframe to a Canonical Dataframe}
\usage{
glucose_df_from_libreview_raw(libreview_raw)
}
\arguments{
\item{libreview_raw}{dataframe of raw values as read from Libreview CSV file}
}
\value{
dataframe in canonical format
}
\description{
Convert a Libreview Dataframe to a Canonical Dataframe
}
