% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrisense_read.R
\name{glucose_df_from_nutrisense}
\alias{glucose_df_from_nutrisense}
\title{read a Nutrisense export file and convert to the PSI CGM format}
\usage{
glucose_df_from_nutrisense(
  filepath = file.path(config::get("tastermonial")$datadir,
    "1627653419114_export-SherriJo.csv"),
  user_id = 1005,
  tz = Sys.timezone()
)
}
\arguments{
\item{filepath}{path to a valid Nutrisense data file}

\item{user_id}{new user ID to be appended to the dataframe}

\item{tz}{time zone}
}
\value{
a canonical glucose value dataframe
}
\description{
Nutrisense export files contain timestamped information about
glucose values, meals, exercise, sleep, and more.
Important: This function expects an extra line at the top of the Nutrisense raw file.
The first line must say "Nutrisense" and include the full name of the person.
}
