% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{glucose_for_food_df}
\alias{glucose_for_food_df}
\title{return a dataframe of rows in the database where food matches food}
\usage{
glucose_for_food_df(
  conn_args = config::get("dataconnection"),
  user_id = 1235,
  foodname = "blueberries"
)
}
\arguments{
\item{conn_args}{database connection}

\item{user_id}{user ID}

\item{foodname}{a string indicating a specific food}
}
\value{
a valid glucose dataframe containing records matching \code{food}
}
\description{
Search the notes database for records indicating \code{foodname} and
return just those rows that contain that food.
}
