% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glucose_raw}
\alias{glucose_raw}
\title{Glucose data}
\format{
a dataframe of 6 variables: time, scan, hist, strip, food, value
}
\usage{
glucose_raw
}
\description{
Glucose data
}
\keyword{datasets}
