% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm_display.R
\docType{data}
\name{glucose_target_gg}
\alias{glucose_target_gg}
\title{glucose_target_gg}
\format{
An object of class \code{LayerInstance} (inherits from \code{Layer}, \code{ggproto}, \code{gg}) of length 11.
}
\usage{
glucose_target_gg
}
\description{
a handy ggplot object that draws a band through the "healthy" target zones across the width of any graph:
}
\keyword{datasets}
