% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_times_calc.R
\name{interpolated}
\alias{interpolated}
\title{(UNFINISHED) Interpolate values between two vectors}
\usage{
interpolated(
  v1,
  v2,
  operator = function(x) {
     x
 }
)
}
\arguments{
\item{v1}{first vector}

\item{v2}{second vector}

\item{operator}{any function}
}
\value{
vector
}
\description{
Given two vectors \code{v1} and \code{v2},
return a new vector that has the same length as base
}
