% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{libreview_csv_df}
\alias{libreview_csv_df}
\title{Read a standard format Libreview CSV file}
\usage{
libreview_csv_df(
  file = system.file("extdata", package = "cgmr", "Firstname2Lastname2_glucose.csv")
)
}
\arguments{
\item{file}{path to a valid Libreview CSV file}
}
\value{
a canonical glucose value dataframe
}
\description{
Read a raw libreview CSV file and return it as a plain dataframe
}
