% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{make_zero_time_df}
\alias{make_zero_time_df}
\title{Normalize time dataframe to zero}
\usage{
make_zero_time_df(df)
}
\arguments{
\item{df}{valid glucose dataframe}
}
\value{
dataframe
}
\description{
Normalize time dataframe to zero
}
