% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_calc.R
\name{match_food}
\alias{match_food}
\title{Timestamps for specific foods}
\usage{
match_food(foodlist, activity_df)
}
\arguments{
\item{foodlist}{character vector of names of food}

\item{activity_df}{activity dataframe}
}
\description{
dataframe of timestamps and names of food
}
