% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{max_date_for_user}
\alias{max_date_for_user}
\title{Most recent date in the database for a given user}
\usage{
max_date_for_user(
  conn_args = config::get("dataconnection"),
  user_id = 1234,
  fromDate = "2019-11-01",
  table_name = conn_args$glucose_table
)
}
\arguments{
\item{conn_args}{valid database connection}

\item{user_id}{user ID}

\item{fromDate}{data object representing the first date to return}

\item{table_name}{the table in which to find the latest record (currently fixed at \code{glucose-records})}
}
\value{
a date object representing the most recent record in the database for this user. NA if there are no records.
}
\description{
Search the default database for the most recent (aka latest) timestamp
Note: doesn't currently work for tables other than \code{glucose_records}
}
