% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{name_from_libreview_file}
\alias{name_from_libreview_file}
\title{name of the person associated with a Libreview glucose file.}
\usage{
name_from_libreview_file(filepath)
}
\arguments{
\item{filepath}{path to the CSV file}
}
\value{
a space-separated character string made of first_name last_name. NA if no name found
}
\description{
Given a valid Libreview file, return a string of the form first_name last_name
}
