% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{normalize_to_zero}
\alias{normalize_to_zero}
\title{normalize all \code{df$value} to a range from 0 to 1}
\usage{
normalize_to_zero(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
dataframe
}
\description{
Assumes \verb{df includes columns }t\code{and}value`
}
