% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_util.R
\name{normalize_value}
\alias{normalize_value}
\title{normalize with prefixlength}
\usage{
normalize_value(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
dataframe
}
\description{
Given any dataframe \code{df} with
columns \code{t} and \code{value}, return an equivalent dataframe where all the \code{value}s
are reset such that they start at 0 instead of whatever they started at.
}
