% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{notes_df_from_csv}
\alias{notes_df_from_csv}
\title{Notes dataframe from a CSV}
\usage{
notes_df_from_csv(
  file = system.file("extdata", package = "cgmr", "FirstName1Lastname1_notes.csv"),
  user_id = 1235
)
}
\arguments{
\item{file}{path to a valid notes CSV file}

\item{user_id}{user ID to associate with this dataframe}
}
\value{
dataframe for a valid notes CSV file
}
\description{
Return a canonical notes dataframe from a properly-constructed Notes CSV file.
}
