% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{notes_df_from_glucose_table}
\alias{notes_df_from_glucose_table}
\title{Return all \code{glucose_records} that have something in the notes field}
\usage{
notes_df_from_glucose_table(
  conn_args = config::get("dataconnection"),
  user_id = 1234
)
}
\arguments{
\item{conn_args}{valid database connection with a \code{glucose_records} table}

\item{user_id}{User ID}
}
\value{
dataframe of
}
\description{
Return all \code{glucose_records} that have something in the notes field
}
