% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{notes_df_from_notes_table}
\alias{notes_df_from_notes_table}
\title{Read notes dataframe from database}
\usage{
notes_df_from_notes_table(
  conn_args = config::get("dataconnection"),
  user_id = 1235,
  fromDate = "2019-11-01",
  db_filter = NULL
)
}
\arguments{
\item{conn_args}{valid database connection}

\item{user_id}{user id}

\item{fromDate}{(optional) earliest date from which to return notes}

\item{db_filter}{A function for filtering the database. Use instead of \code{from_date} or \code{user_id}}
}
\value{
dataframe representation of all notes for that user
}
\description{
If notes exist for ID, return all notes in a dataframe
}
