% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrisense_read.R
\name{nutrisense_raw}
\alias{nutrisense_raw}
\title{Raw Dataframe from Nutrisense Filepath}
\usage{
nutrisense_raw(
  filepath = system.file("extdata", package = "cgmr",
    "Firstname3Lastname3_nutrisense.csv"),
  skip_lines = 1
)
}
\arguments{
\item{filepath}{path to a Nutrisense CSV file}

\item{skip_lines}{integer lines to skip while reading}
}
\value{
dataframe raw nutrisense data format
}
\description{
Raw Dataframe from Nutrisense Filepath
}
