% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{old_food_times_df}
\alias{old_food_times_df}
\title{Glucose values after eating a specific food}
\usage{
old_food_times_df(user_id = 1235, timeLength = 120, foodname = "watermelon")
}
\arguments{
\item{user_id}{user ID}

\item{timeLength}{number of minutes for the glucose record to show after the food was eaten.}

\item{foodname}{character string representing the food item of interest}
}
\value{
dataframe
}
\description{
return a dataframe of the Glucose values for a \code{timeLength}
following \code{foodname} appearing in \code{notes_records}
}
