% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{old_normalize_value}
\alias{old_normalize_value}
\title{return a new df where value are normalized to start from zero.}
\usage{
old_normalize_value(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
Useful when transforming \code{food_times_df}, this will reset all time values
back to offsets from the earliest time in the dataframe.
Assumes a dataframe with column \code{value}.
}
