% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_by_meal.R
\name{plot_food}
\alias{plot_food}
\title{Plot Glucose Level Changes Over Time}
\usage{
plot_food(foodresults = NULL, foodname = "Food")
}
\arguments{
\item{foodresults}{A dataframe, usually the result of a call to \link{food_times_df}.
The dataframe should have the following columns:
\itemize{
\item t: a numeric vector representing time.
\item value: a numeric vector representing glucose level changes.
\item meal: a character vector representing meal names.
If NULL, the function will return NULL. Default is NULL.
}}

\item{foodname}{a string representing the name of the food to plot. Note: this is for display only; it does not filter or otherwise generate different results.}
}
\value{
A ggplot object representing the line plot of glucose level changes over time.
}
\description{
This function creates a line plot of glucose level changes over time. The changes are colored by meals. The function includes a grey rectangle behind the line plot for contrast.
}
