% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_by_meal.R
\name{plot_intervention}
\alias{plot_intervention}
\title{Plot Intervention Over Time}
\usage{
plot_intervention(foodresults = NULL, intervention = "glucotrojan")
}
\arguments{
\item{foodresults}{A data frame containing the glucose measurements. This is typically generated
by calling \code{\link{food_times_df}} with a valid object created by \code{\link{cgm_data}}
and a 'foodname' string. The dataframe should contain columns named 't', 'value', 'meal', and 'control'.
't' should be the time measurements.
'value' should be the glucose measurements.
'meal' identifies unique meals.
'control' is a logical indicating whether an intervention was applied.
If this parameter is NULL, the function will return NULL.}

\item{intervention}{A string representing the intervention (e.g., "glucotrojan").
The function will check whether each meal includes this intervention.}
}
\value{
A ggplot2 object representing the glucose response plot. The plot can be further customized
using additional ggplot2 functions.
}
\description{
This function generates a plot of glucose responses over time, showing
vertical bars representing the range of glucose values for each time point.
The bars are color-coded based on whether the meal includes the specified intervention.
}
