% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{table_df_from_db}
\alias{table_df_from_db}
\title{Load all rows from a database table and return as a dataframe}
\usage{
table_df_from_db(
  conn_args = config::get("dataconnection"),
  table_name = "glucose_records"
)
}
\arguments{
\item{conn_args}{valid database connection}

\item{table_name}{character string}
}
\value{
dataframe representation of the table
}
\description{
Connects to the default database and looks up a table.
Slightly more convenient than setting up the database connection first.
}
