% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrisense_read.R
\name{xform_nutrisense}
\alias{xform_nutrisense}
\title{Transform Raw Nutrisense into Canonical Glucose Format}
\usage{
xform_nutrisense(raw_data, tz)
}
\arguments{
\item{raw_data}{dataframe pulled from the Nutrisense file}

\item{tz}{character string for time zone}
}
\value{
dataframe canonical Glucose format
}
\description{
Transform Raw Nutrisense into Canonical Glucose Format
}
