% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{zero_time}
\alias{zero_time}
\title{Convert the timestamp into time objects}
\usage{
zero_time(times_vector)
}
\arguments{
\item{times_vector}{vector of start times.}
}
\value{
a vector normalized to the beginning of the sequence
}
\description{
Convert the timestamp into time objects
}
