% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_util.R
\name{df_for_all_auc}
\alias{df_for_all_auc}
\title{make one dataframe showing \code{auc_for_food()} for a vector of food names}
\usage{
df_for_all_auc(food_list, cgm_data)
}
\arguments{
\item{food_list}{vector of character strings with each foodname}

\item{cgm_data}{CGM data object}
}
\value{
dataframe representing all results of auc_for_food() calls
}
\description{
make one dataframe showing \code{auc_for_food()} for a vector of food names
}
