% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_utils.R
\name{food_times_df}
\alias{food_times_df}
\title{Glucose values after eating a specific food}
\usage{
food_times_df(
  conn_args = config::get("dataconnection"),
  user_id = NULL,
  timeLength = 120,
  prefixLength = 0,
  foodname = "watermelon",
  db_filter = function(x) {     x }
)
}
\arguments{
\item{conn_args}{database connection}

\item{user_id}{user ID}

\item{timeLength}{number of minutes for the glucose record to show after the food was eaten.}

\item{prefixLength}{number of additional minutes to add before the starttime.}

\item{foodname}{character string representing the food item of interest}
}
\value{
dataframe
}
\description{
return a dataframe of the Glucose values for a \code{timeLength}
following \code{foodname} appearing in \code{notes_records}.
This function calls the database directly
and is intended to work standalone, without other functions.
}
