% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_util.R
\name{normalize_value}
\alias{normalize_value}
\title{normalize with prefixlength}
\usage{
normalize_value(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
dataframe
}
\description{
Assumes \verb{df includes columns }t\code{and}value`
}
