% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\docType{data}
\name{NOTES_COLUMNS}
\alias{NOTES_COLUMNS}
\title{Possible values for \code{Activity} column in Notes.}
\format{
An object of class \code{character} of length 7.
}
\usage{
NOTES_COLUMNS
}
\description{
Possible values for \code{Activity} column in Notes.
}
\keyword{datasets}
