% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{notes_df_from_glucose_table}
\alias{notes_df_from_glucose_table}
\title{Return all \code{glucose_records} that have something in the notes field}
\usage{
notes_df_from_glucose_table(glucose_records, user_id = 1234)
}
\arguments{
\item{glucose_records}{valid glucose record dataframe}

\item{user_id}{User ID}
}
\value{
dataframe of
}
\description{
Notes are sometimes located in the "Notes" field of a CSV file,
which in turn will be saved in \code{glucose_records}.
Harvest any such notes and place them in \code{notes_records}.
}
